+ try on no file length + deny to download not ready files
