+++ test/uk/gov/hmrc/fileupload/controllers/FileControllerSpec.scala
@@ -31,7 +31,7 @@ import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import uk.gov.hmrc.fileupload._
 import uk.gov.hmrc.fileupload.infrastructure.{AlwaysAuthorisedBasicAuth, BasicAuth}
-import uk.gov.hmrc.fileupload.read.envelope.{File, FileStatusQuarantined, WithValidEnvelope}
+import uk.gov.hmrc.fileupload.read.envelope.{File, FileStatusAvailable, WithValidEnvelope}
 import uk.gov.hmrc.fileupload.write.envelope.EnvelopeCommand
 import uk.gov.hmrc.fileupload.write.infrastructure.{CommandAccepted, CommandNotAccepted}
 import uk.gov.hmrc.play.test.UnitSpec
@@ -60,7 +60,7 @@ class FileControllerSpec extends UnitSpec with ScalaFutures {
   val envelopeId = EnvelopeId()
   val fileId = FileId()
   val refId = FileRefId("someTextNotUuidFormat")
-  val file = File(fileId, refId, FileStatusQuarantined, name = Some("myfile.txt"), length = Some(100))
+  val file = File(fileId, refId, FileStatusAvailable, name = Some("myfile.txt"), length = Some(100))
   val envelope = Support.envelope.copy(files = Some(Seq(file))).copy(_id = envelopeId)
   val source = Source.empty[ByteString]
   val authHeaders = HeaderNames.AUTHORIZATION -> ("Basic " + basic64("yuan:yaunspassword"))
